requirements.txt
Django==5.1.1
djangorestframework==3.15.2
psycopg2-binary==2.9.9
Pillow==10.4.0
pytest==8.3.3
pytest-django==4.8.0
coverage==7.6.1
ruff==0.6.9
black==24.8.0
django-cors-headers==4.9.0
djangorestframework
django-cors-headers
django-prometheus